requirements.txt
django
djangorestframework
pyserial
pytest
